move utils in own folder
